MNT: Rename SDS-in-a-box to sds-data-manager
Renaming the directories within git as well.
