+++ .devcontainer/Dockerfile
@@ -10,6 +10,6 @@ RUN curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2
 RUN unzip awscliv2.zip
 RUN sudo ./aws/install
 RUN sudo apt-get install -y python3 python3-pip
-ENV AWS_CONFIG_FILE=/workspaces/SDS-in-a-box/.aws/config
-ENV AWS_SHARED_CREDENTIALS_FILE=/workspaces/SDS-in-a-box/.aws/credentials
-WORKDIR /workspaces/SDS-in-a-box/
+ENV AWS_CONFIG_FILE=/workspaces/sds-data-manager/.aws/config
+ENV AWS_SHARED_CREDENTIALS_FILE=/workspaces/sds-data-manager/.aws/credentials
+WORKDIR /workspaces/sds-data-manager/
