+++ app.py
@@ -4,7 +4,7 @@ import string
 
 import aws_cdk as cdk
 
-from sds_in_a_box.sds_in_a_box_stack import SdsInABoxStack
+from sds_data_manager.sds_data_manager_stack import SdsDataManagerStack
 
 app = cdk.App()
 
@@ -19,5 +19,5 @@ elif SDS_ID == "random":
     SDS_ID = "".join([random.choice(string.ascii_lowercase) for i in range(8)])
 
 
-SdsInABoxStack(app, f"SdsInABoxStack-{SDS_ID}", SDS_ID)
+SdsDataManagerStack(app, f"SdsDataManagerStack-{SDS_ID}", SDS_ID)
 app.synth()
