+++ tests/unit/opensearch_utils/test_action.py
@@ -1,6 +1,6 @@
 import unittest
 
-from sds_in_a_box.SDSCode.opensearch_utils.action import Action
+from sds_data_manager.SDSCode.opensearch_utils.action import Action
 
 
 class TestAction(unittest.TestCase):
