+++ tests/unit/opensearch_utils/test_client.py
@@ -6,12 +6,12 @@ import pytest
 from botocore.exceptions import ClientError
 from opensearchpy import RequestsHttpConnection
 
-from sds_in_a_box.SDSCode.opensearch_utils.action import Action
-from sds_in_a_box.SDSCode.opensearch_utils.client import Client
-from sds_in_a_box.SDSCode.opensearch_utils.document import Document
-from sds_in_a_box.SDSCode.opensearch_utils.index import Index
-from sds_in_a_box.SDSCode.opensearch_utils.payload import Payload
-from sds_in_a_box.SDSCode.opensearch_utils.query import Query
+from sds_data_manager.SDSCode.opensearch_utils.action import Action
+from sds_data_manager.SDSCode.opensearch_utils.client import Client
+from sds_data_manager.SDSCode.opensearch_utils.document import Document
+from sds_data_manager.SDSCode.opensearch_utils.index import Index
+from sds_data_manager.SDSCode.opensearch_utils.payload import Payload
+from sds_data_manager.SDSCode.opensearch_utils.query import Query
 
 
 @pytest.mark.network()
