+++ tests/unit/opensearch_utils/test_index.py
@@ -1,6 +1,6 @@
 import unittest
 
-from sds_in_a_box.SDSCode.opensearch_utils.index import Index
+from sds_data_manager.SDSCode.opensearch_utils.index import Index
 
 
 class TestIndex(unittest.TestCase):
