+++ tests/unit/opensearch_utils/test_payload.py
@@ -1,9 +1,9 @@
 import unittest
 
-from sds_in_a_box.SDSCode.opensearch_utils.action import Action
-from sds_in_a_box.SDSCode.opensearch_utils.document import Document
-from sds_in_a_box.SDSCode.opensearch_utils.index import Index
-from sds_in_a_box.SDSCode.opensearch_utils.payload import Payload
+from sds_data_manager.SDSCode.opensearch_utils.action import Action
+from sds_data_manager.SDSCode.opensearch_utils.document import Document
+from sds_data_manager.SDSCode.opensearch_utils.index import Index
+from sds_data_manager.SDSCode.opensearch_utils.payload import Payload
 
 
 class TestPayload(unittest.TestCase):
